refactor: rename `formatDate` filename to dash-case and move to utils
